requirements.txt
requests==2.31.0
mysql-connector-python==8.2.0
numpy==1.24.3
pandas==2.0.3
pandas-ta==0.3.14b0p
ython-telegram-bot==20.7
python-dotenv==1.0.0
flask==3.0.0